refactor(components): rename tags to layout #9
