+++ src/styles/all.sass
@@ -22,11 +22,11 @@
 @import 'menu/menu-item'
 @import 'menu/menu-trigger'
 
-// tags first; pages can customize
-@import 'tags/header'
-@import 'tags/layout'
-@import 'tags/page'
-@import 'tags/card'
+// layout first; pages can customize
+@import 'layout/header'
+@import 'layout/layout'
+@import 'layout/page'
+@import 'layout/card'
 
 @import 'pages/article'
 @import 'pages/home'
